requirements.txt
boto==2.45.0
cffi==1.9.1
click==6.6
cryptography==1.7.1
enum34==1.1.6
idna==2.2
ipaddress==1.0.17
pyasn1==0.1.9
pycparser==2.17
pyperclip==1.5.27
PyYAML==3.12
six==1.10.0
